+++ src/alias/scn.js
@@ -1,0 +1,5 @@
+#!/usr/bin/env node
+
+import { runAlias } from '../helpers/cmds.js'
+
+runAlias('swpm', ['clean', '--node-modules'])