+++ CHANGELOG.md
@@ -15,6 +15,7 @@ Security    to invite users to upgrade in case of vulnerabilities.
 
 - `si` alias for `swpm install`
 - `sa` alias for `swpm add`
+- multiple alias for `swpm clean`
 
 ### Changed