+++ CHANGELOG.md
@@ -15,6 +15,9 @@ Security    to invite users to upgrade in case of vulnerabilities.
 
 - `si` alias for `swpm install`
 - `sa` alias for `swpm add`
+- `sae` alias for `swpm add <package> --save-exact`
+- `sad` alias for `swpm add <package> --save-dev`
+- `sade` alias for `swpm add <package> --save-dev --save-exact`
 - multiple alias for `swpm clean`
 - `sx` alias for `swpx`